+++ CHANGELOG.md
@@ -17,7 +17,7 @@ All notable changes to this project will be documented in this file.
 - The `/etc/pacman.d/mirrors` dir has been removed.
   - All data files now exist in `/var/lib/pacman-mirrors`.
   - If the `Custom` mirrorfile exist it will convert to `custom-mirrors.json`
-- A lot of inevitable small fixes
+- A lot of inevitable small fixes.
 
 ## [3.2.2] - 2017-02-12
 - Fix issue with multiple country select.